Moving source folder inside viz folder
