+++ src/lib/dataviews/DataViewWrapper.ts
@@ -1,6 +1,6 @@
 import { WithEvents } from '@/core/mixins/WithEvents';
 import { Layer } from '@/viz';
-import { Source } from '@/source';
+import { Source } from '@/viz/source';
 import { Filter } from '@/viz/filters/types';
 import { AggregationType } from '@/data/operations/aggregation/aggregation';
 import { DataViewImpl } from './DataViewImpl';
