+++ src/lib/dataviews/formula/Formula.ts
@@ -1,5 +1,5 @@
 import { Layer } from '@/viz';
-import { Source } from '@/source';
+import { Source } from '@/viz/source';
 import { BuiltInFilters } from '@/viz/filters/types';
 import { uuidv4 } from '@/core/utils/uuid';
 import { DataViewCalculation } from '../mode/DataViewMode';
