+++ src/lib/viz/__tests__/styles/SizeContinuousStyle.spec.ts
@@ -1,5 +1,5 @@
 import { Deck } from '@deck.gl/core';
-import { DatasetSource } from '@/source';
+import { DatasetSource } from '@/viz/source';
 import { sizeContinuousStyle } from '../../style';
 import { getDefaultSizeRange } from '../../style/helpers/size-continuous-style';
 import * as mapsResponse from '../data-mocks/maps.number.json';
@@ -20,7 +20,7 @@ const getMetadata = jest.fn().mockImplementation(() => {
   };
 });
 
-jest.mock('../../../source/DatasetSource', () => ({
+jest.mock('../../source/DatasetSource', () => ({
   DatasetSource: jest.fn().mockImplementation(() => ({ getMetadata }))
 }));
