+++ src/lib/viz/source/DOSource.ts
@@ -1,7 +1,7 @@
 import { Credentials, defaultCredentials } from '@/auth';
 import { Source, SourceProps, SourceMetadata, NumericFieldStats } from './Source';
-import { parseGeometryType } from '../viz/style/helpers/utils';
-import { sourceErrorTypes, SourceError } from '../viz/errors/source-error';
+import { parseGeometryType } from '../style/helpers/utils';
+import { sourceErrorTypes, SourceError } from '../errors/source-error';
 
 /* eslint-disable camelcase */
 interface Variable {
