+++ src/lib/viz/source/GeoJSONSource.ts
@@ -13,8 +13,8 @@ import {
   shouldInitialize
 } from './Source';
 
-import { sourceErrorTypes, SourceError } from '../viz/errors/source-error';
-import { selectPropertiesFrom } from '../viz/utils/object';
+import { sourceErrorTypes, SourceError } from '../errors/source-error';
+import { selectPropertiesFrom } from '../utils/object';
 
 interface GeoJSONSourceProps extends SourceProps {
   data: GeoJSON;
