+++ src/lib/viz/style/helpers/size-continuous-style.ts
@@ -1,4 +1,4 @@
-import { NumericFieldStats, GeometryType } from '@/source';
+import { NumericFieldStats, GeometryType } from '@/viz/source';
 import { StyledLayer } from '../layer-style';
 import { range } from './math-utils';
 import { CartoStylingError, stylingErrorTypes } from '../../errors/styling-error';
