+++ src/lib/viz/style/helpers/utils.ts
@@ -1,6 +1,6 @@
 import { RGBAColor } from '@deck.gl/core';
 import { CartoError } from '@/core/errors/CartoError';
-import { GeometryType } from '@/source';
+import { GeometryType } from '@/viz/source';
 import { getColorPalette } from '../palettes';
 import { Classifier } from '../../utils/Classifier';
