+++ src/lib/viz/style/layer-style.ts
@@ -1,5 +1,5 @@
 import { Deck } from '@deck.gl/core';
-import { Source } from '@/source';
+import { Source } from '@/viz/source';
 
 export interface StyledLayer {
   getMapInstance(): Deck;
